pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "fair-manufacturing-recsys"
version = "0.1.0"
description = "Fair Manufacturing Service Recommendation — RecSys 2026"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torch-geometric>=2.4.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "black>=23.0", "ruff>=0.1.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
